msm: sde: workaround to fix dual display issue on boot
Issue:

With HDMI and TypeC DP connected on boot and power on the System,
nothing can be seen on both display and following errors are
seen in the console:

[sde error]invalid coordinates, stage:0 l:0-2560 r:0-1920

Fix:

Inspired by:
https://patches.linaro.org/project/linux-arm-msm/patch/20220209172520.3719906-18-dmitry.baryshkov@linaro.org/#732400

Comment:

Qualcom has no support to dual display and customers need to implement
on their own.

This is a workaround fix without rootcause identified. Need to look
at it sometime later.